change the way we handle non-inlined jump tables.
The new approach gets rid of our changes/hacks to ARMConstantPoolValue.cpp and uses a more standard approach:
We use movt/movw pairs to materialize the address which gets rid of one class of constant pool entries.

if movt/movw is not available we will likely get and assert in one of the lowering phases but we do not support older ARM architectures in pnacl anyway.