This CL adds a new flag "-add-libgcc-dependencies" to llc mimicking --add-nacl-read-tp-dependency. Also, slightly over commented this and previous changes
Ultimately, this results in all function of libgcc to be pulled into the pexe.

This is CL is also a starting point for the discussion of which function should be expected to be part of libgcc.

A full list of the candidates can be found in
llvm-trunk/lib/CodeGen/SelectionDAG/TargetLowering.cpp
look for InitLibcallNames()

To test/try this change:
* build bitcode libgcc and removed native versions
tools/llvm/utman.sh build-compiler-rt-bitcode
* patch the driver http://codereview.chromium.org/7236001
* update live driver
tools/llvm/utman.sh driver
* run some tests
./scons bitcode=1 platform=arm -j 8 -k



BUG= http://code.google.com/p/nativeclient/issues/detail?id=842
BUG= http://code.google.com/p/nativeclient/issues/detail?id=1520


CL= http://codereview.chromium.org/7233016/